+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 - Change file size math and add current size to output
+- Strip whitespace when checking broken internal links
 
 ## [2.1.0] - 2021-01-22